Update account.go
This change adds a few more error checking, and also attempts to close the response body only if not nil. 
There were a few instances, where the development laptop went into sleep mode then, when it went back to normal operation, the dns resolver was not operational for a few seconds.
This triggered a panic because the error had not been previously captured.